pyproject.toml
[tool.poetry]
name = "number-detection"
version = "0.1.0"
description = ""
authors = ["Abhinav Tushar <abhinav@lepisma.xyz>"]
readme = "README.org"

[tool.poetry.dependencies]
python = "^3.12"
transformers = "^4.42.4"
docopt = "^0.6.2"
tqdm = "^4.66.4"
pandas = "^2.2.2"
word2number = "^1.1"
werpy = "^2.1.2"
deepgram-sdk = ">=3,<4.0"
google-cloud-speech = "^2.26.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
